+++ src/TemperatureDataBaseWPF/ViewModels/Pages/SetParametersPageViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace TemperatureDataBaseWPF.ViewModels.Pages
+{
+    public class SetParametersPageViewModel : ObservableObject
+    {
+
+    }
+}