+++ src/TemperatureDataBaseWPF/ViewModels/Windows/MainWindowViewModel.cs
@@ -27,9 +27,9 @@ namespace TemperatureDataBaseWPF.ViewModels.Windows
             },
             new NavigationViewItem()
             {
-                Content = "Data",
-                Icon = new SymbolIcon { Symbol = SymbolRegular.DataHistogram24 },
-                TargetPageType = typeof(Views.Pages.DataPage)
+                Content = "Parameters",
+                Icon = new SymbolIcon { Symbol = SymbolRegular.EditSettings24 },
+                TargetPageType = typeof(Views.Pages.SetParametersPageView)
             }
         };