+++ CHANGELOG.md
@@ -26,6 +26,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - 2023-11-09
 
+### ADDED
+- Set parameters view and view model.
+- Set parameters added in dependece injection.
+
+---
+
+## [Unreleased] - 2023-11-09
+
 ### FIXED
 - Updates the gauge only if the temperature value is less than 100.